fix: address PRRT_kwDOSJAM6s6HqxnT — split oversized pr_monitor test part under 1500-line guardrail
test_pr_monitor_part_003.py had grown to 1,581 lines via subsequent CI-rerun
fix commits, tripping MAX_FIRST_PARTY_FILE_LINES=1_500 in
test_first_party_code_files_stay_under_line_limit. Move the CI-failure rerun
budget / state-key and misc failure tests into a new
test_pr_monitor_part_004.py, dropping part_003 to 1,350 lines.
