+++ tests/unit/runtime/test_pr_monitor_parts/test_pr_monitor_part_003.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 import pytest
 
 from awf.runtime.pr_monitor import (
-    AddressComments,
     CheckFailure,
     CheckState,
     MergeableState,
@@ -21,9 +20,7 @@ from awf.runtime.pr_monitor import (
     RerunTransientCI,
     ReviewComment,
     ReviewThread,
-    _ci_transient_rerun_state_key,
     _log_shows_docker_registry_timeout,
-    _should_rerun_transient_ci,
     decide,
 )
 
@@ -1348,234 +1345,3 @@ class TestCiFailure:
 
         assert isinstance(action, ReportCiFailure)
         assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_rerun_state_key_uses_structured_failure_signature(self) -> None:
-        left = (
-            CheckFailure(
-                name="lint:type",
-                conclusion="FAILURE",
-                log_excerpt="HTTP 502",
-                run_id="run",
-            ),
-        )
-        right = (
-            CheckFailure(
-                name="type",
-                conclusion="FAILURE",
-                log_excerpt="HTTP 502",
-                run_id="run:lint",
-            ),
-        )
-
-        assert _ci_transient_rerun_state_key("head", left) != _ci_transient_rerun_state_key(
-            "head",
-            right,
-        )
-
-    @pytest.mark.unit
-    def test_transient_failure_falls_back_to_agent_after_rerun_budget(self) -> None:
-        failure = CheckFailure(
-            name="CI",
-            conclusion="FAILURE",
-            log_excerpt="curl: (56) Recv failure: Connection reset by peer",
-            run_id="25655330295",
-        )
-        status = _status(check_state=CheckState.FAILURE, ci_failures=(failure,))
-        state = MonitorState()
-        state.threads_addressed_ids[
-            _ci_transient_rerun_state_key(status.head_sha, status.ci_failures)
-        ] = "2"
-
-        action = decide(status, state, MonitorConfig(ci_transient_rerun_max_attempts=2))
-
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_transient_rerun_budget_reads_legacy_rollup_signature(self) -> None:
-        failure = CheckFailure(
-            name="python-full-coverage",
-            conclusion="FAILURE",
-            log_excerpt="HTTP status server error (502 Bad Gateway)",
-            run_id="25655330295",
-        )
-        rollup_failure = CheckFailure(
-            name="ci-required",
-            conclusion="FAILURE",
-            log_excerpt="A required CI job did not pass.",
-            run_id="25655330295",
-        )
-        status = _status(
-            check_state=CheckState.FAILURE,
-            ci_failures=(failure, rollup_failure),
-        )
-        state = MonitorState()
-        state.threads_addressed_ids[
-            _ci_transient_rerun_state_key(status.head_sha, status.ci_failures)
-        ] = "2"
-
-        action = decide(status, state, MonitorConfig(ci_transient_rerun_max_attempts=2))
-
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == status.ci_failures
-
-    @pytest.mark.unit
-    def test_transient_failure_with_disabled_rerun_budget_dispatches_agent_repair(self) -> None:
-        failure = CheckFailure(
-            name="CI",
-            conclusion="FAILURE",
-            log_excerpt="HTTP status server error (503 Service Unavailable)",
-            run_id="25655330295",
-        )
-
-        action = decide(
-            _status(check_state=CheckState.FAILURE, ci_failures=(failure,)),
-            MonitorState(),
-            MonitorConfig(ci_transient_rerun_max_attempts=0),
-        )
-
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_transient_failure_corrupt_rerun_count_is_treated_as_zero(self) -> None:
-        failure = CheckFailure(
-            name="CI",
-            conclusion="FAILURE",
-            log_excerpt="curl: (56) Recv failure: Connection reset by peer",
-            run_id="25655330295",
-        )
-        status = _status(check_state=CheckState.FAILURE, ci_failures=(failure,))
-        state = MonitorState()
-        state.threads_addressed_ids[
-            _ci_transient_rerun_state_key(status.head_sha, status.ci_failures)
-        ] = "not-an-int"
-
-        action = decide(status, state, MonitorConfig())
-
-        assert isinstance(action, RerunTransientCI)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_infra_assert_log_does_not_mask_transient_ci_rerun(self) -> None:
-        failure = CheckFailure(
-            name="CI",
-            conclusion="FAILURE",
-            log_excerpt="assert passed while reconnecting runner\nHTTP 502 Bad Gateway",
-            run_id="25655330295",
-        )
-
-        action = decide(
-            _status(check_state=CheckState.FAILURE, ci_failures=(failure,)),
-            MonitorState(),
-            MonitorConfig(),
-        )
-
-        assert isinstance(action, RerunTransientCI)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_structured_test_evidence_prevents_transient_ci_rerun(self) -> None:
-        failure = CheckFailure(
-            name="python-full-coverage",
-            conclusion="FAILURE",
-            log_excerpt="curl: (56) Recv failure: Connection reset by peer",
-            run_id="25655330295",
-            test_node_ids=("pkg/tests/test_api.py::test_x",),
-            assertion_snippets=("E   AssertionError: boom",),
-        )
-
-        action = decide(
-            _status(check_state=CheckState.FAILURE, ci_failures=(failure,)),
-            MonitorState(),
-            MonitorConfig(),
-        )
-
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_transient_rerun_helper_rejects_empty_failure_snapshot(self) -> None:
-        assert not _should_rerun_transient_ci(
-            _status(check_state=CheckState.FAILURE, ci_failures=()),
-            MonitorState(),
-            MonitorConfig(),
-        )
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("run_id", [None, ""])
-    def test_transient_failure_without_run_id_dispatches_agent_repair(
-        self,
-        run_id: str | None,
-    ) -> None:
-        failure = CheckFailure(
-            name="CI",
-            conclusion="FAILURE",
-            log_excerpt="HTTP status server error (502 Bad Gateway)",
-            run_id=run_id,
-        )
-
-        action = decide(
-            _status(check_state=CheckState.FAILURE, ci_failures=(failure,)),
-            MonitorState(),
-            MonitorConfig(),
-        )
-
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_code_like_failure_still_dispatches_agent_repair(self) -> None:
-        failure = CheckFailure(
-            name="python-full-coverage",
-            conclusion="FAILURE",
-            log_excerpt="FAILED tests/unit/test_thing.py::test_case - AssertionError",
-            run_id="25655330295",
-        )
-
-        action = decide(
-            _status(check_state=CheckState.FAILURE, ci_failures=(failure,)),
-            MonitorState(),
-            MonitorConfig(),
-        )
-
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_failure_with_per_check_details(self) -> None:
-        failure = CheckFailure(
-            name="playwright", conclusion="FAILURE", log_excerpt="Error: timeout"
-        )
-        action = decide(
-            _status(check_state=CheckState.FAILURE, ci_failures=(failure,)),
-            MonitorState(),
-            MonitorConfig(),
-        )
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == (failure,)
-
-    @pytest.mark.unit
-    def test_failure_with_empty_failure_list_still_returns_action(self) -> None:
-        """Runner can still fetch logs on its own via ``gh run view``."""
-        action = decide(
-            _status(check_state=CheckState.FAILURE),
-            MonitorState(),
-            MonitorConfig(),
-        )
-        assert isinstance(action, ReportCiFailure)
-        assert action.failures == ()
-
-    @pytest.mark.unit
-    def test_unresolved_comments_take_priority_over_ci_failure(self) -> None:
-        """A new comment arriving mid-CI-fail means: address the comment
-        first; the next push triggers CI anew. Otherwise we'd keep
-        retrying CI fixes for stale commits."""
-        t = _thread()
-        action = decide(
-            _status(check_state=CheckState.FAILURE, inline=(t,)),
-            MonitorState(),
-            MonitorConfig(),
-        )
-        assert isinstance(action, AddressComments)
